Move logger package to common
